change: NodeからDBへの接続をモジュール化
